+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * [UIBULKED-612](https://folio-org.atlassian.net/browse/UIBULKED-612) Preview of errors&warnings is re-rendered in the process of uploading from time to time.
 * [UIBULKED-607](https://folio-org.atlassian.net/browse/UIBULKED-607) Migrate to shared GA workflows.
 * [UIBULKED-617](https://folio-org.atlassian.net/browse/UIBULKED-617) The Build query button remains disabled for some languages.
+* [UIBULKED-604](https://folio-org.atlassian.net/browse/UIBULKED-604) Show status from network in error message.
 
 ## [4.2.2](https://github.com/folio-org/ui-bulk-edit/tree/v4.2.2) (2024-11-15)